SIDE-2425: RG - removing extra tests from index.ts which reoccured after merge
