+++ ldk/javascript/examples/self-test-loop/src/selfTest.ts
@@ -19,10 +19,8 @@ import {
   networkWebSocket,
   processStream,
   processQuery,
-  numberInputs,
   streamCursorPosition,
   vaultReadWrite,
-  testNetworkAndListComponents,
   queryDirectory,
   createAndDeleteFile,
   userJWTTest,
@@ -39,6 +37,8 @@ import {
   testBoxInTheBox,
   buttonWhisper,
   linkWhisper,
+  numberInputs,
+  testNetworkAndListComponents,
   listPairWhisperCopyableValue,
   listPairWhisperCopyableLabel,
   simpleFormWhisper,
